pyproject.toml
[tool.poetry]
name = "happyhouse-backend"
version = "0.1.0"
description = ""
authors = ["ethan <ethanminsu8969@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "happyhouse_backend"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.2"
django-environ = "^0.10.0"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
